pyproject.toml
[project]
name = "voice-evaluater"
version = "1.0.0"
description = "Comprehensive Audio Assessment System with Azure Speech SDK, Azure OpenAI, MTI Analysis, and Confidence Scoring"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "azure-cognitiveservices-speech>=1.34.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.36.0",
    "sentence-transformers>=2.2.0",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "protobuf>=6.33.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
audio = [
    "pydub>=0.25.0",
    "ffmpeg-python>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
